Sonar cloud fix
