+++ packages/relay/tests/lib/eth/eth-helpers.ts
@@ -1,11 +1,12 @@
-export function contractResultsByNumberByIndexURL(number: number, index: number) {
+export function contractResultsByNumberByIndexURL(number: number, index: number): string {
   return `contracts/results?block.number=${number}&transaction.index=${index}&limit=100&order=asc`;
 }
 
-export function contractResultsByHashByIndexURL(hash: string, index: number) {
+export function contractResultsByHashByIndexURL(hash: string, index: number): string {
   return `contracts/results?block.hash=${hash}&transaction.index=${index}&limit=100&order=asc`;
 }
 
-export function balancesByAccountIdByTimestampURL(id: string, timestamp?: string) {
-  return `balances?account.id=${id}${timestamp ? `&timestamp=${timestamp}` : ''}`;
+export function balancesByAccountIdByTimestampURL(id: string, timestamp?: string): string {
+  const timestampQuery = timestamp ? `&timestamp=${timestamp}` : '';
+  return `balances?account.id=${id}${timestampQuery}`;
 }
